+++ src/main/java/com/ajou_nice/with_pet/critical_service/dto/CriticalServiceResponse.java
@@ -1,7 +1,7 @@
-package com.ajou_nice.with_pet.admin.model.criticalservice;
+package com.ajou_nice.with_pet.critical_service.dto;
 
 
-import com.ajou_nice.with_pet.domain.entity.CriticalService;
+import com.ajou_nice.with_pet.critical_service.entity.CriticalService;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
