+++ src/main/java/com/ajou_nice/with_pet/security/configuration/SecurityConfig.java
@@ -28,7 +28,8 @@ public class SecurityConfig {
     private final String[] GET_PERMIT_URL = {
             "/api/v1/petsitter/*",
             "/api/v1/show-petsitter",
-            "/api/v1/show-services",
+            "/api/v1/services",
+            "/api/v2/critical-services",
             "/api/v1/reservation",
             "/api/v1/category",
             "/api/v1/dogs/reservation-dogs",
