+++ src/test/java/com/ajou_nice/with_pet/reservations/ReservationTest.java
@@ -2,8 +2,7 @@ package com.ajou_nice.with_pet.reservations;
 
 
 import com.ajou_nice.with_pet.controller.ReservationController;
-import com.ajou_nice.with_pet.domain.dto.reservation.ReservationRequest;
-import com.ajou_nice.with_pet.domain.entity.CriticalService;
+import com.ajou_nice.with_pet.critical_service.entity.CriticalService;
 import com.ajou_nice.with_pet.domain.entity.Dog;
 import com.ajou_nice.with_pet.domain.entity.Party;
 import com.ajou_nice.with_pet.domain.entity.Pay;
@@ -17,7 +16,7 @@ import com.ajou_nice.with_pet.enums.UserRole;
 import com.ajou_nice.with_pet.exception.AppException;
 import com.ajou_nice.with_pet.exception.ErrorCode;
 import com.ajou_nice.with_pet.fixture.Fixture;
-import com.ajou_nice.with_pet.repository.CriticalServiceRepository;
+import com.ajou_nice.with_pet.critical_service.repository.CriticalServiceRepository;
 import com.ajou_nice.with_pet.repository.DogRepository;
 import com.ajou_nice.with_pet.repository.PartyRepository;
 import com.ajou_nice.with_pet.repository.PayRepository;
